pyproject.toml
[project]
name = "documind"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    # Web framework
    "django>=4.2,<5.0",
    "djangorestframework>=3.15",
    "drf-spectacular>=0.27",
    # Configuration
    "django-environ>=0.11",
    # Database — PostgreSQL + vector search
    "psycopg2-binary>=2.9",
    "pgvector>=0.3",
    # Cache + background tasks
    "redis>=5.0",
    "celery>=5.3",
    # LLM providers
    "openai>=1.30",
    "anthropic>=0.28",
    "instructor>=1.3",
    # Agent framework
    "langchain>=0.2",
    "langgraph>=0.1",
    "langsmith>=0.1",
    # Embeddings + keyword search
    "sentence-transformers>=3.0",
    "rank-bm25>=0.2",
    # Document parsing
    "pypdf>=4.0",
    # File storage — S3-compatible (MinIO locally, AWS S3 in production)
    "boto3>=1.34",
    "django-storages>=1.14",
    "python-json-logger>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "ruff>=0.15.5",
]

# ---------------------------------------------------------------------------
# Ruff — linter + formatter
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 88

[tool.ruff.lint]
# E = pycodestyle errors, F = pyflakes, I = isort
select = ["E", "F", "I"]
ignore = ["E501"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["tests"]
addopts = "-v"
